# Remove usage of `stopifnot()` and `structure()` in `new_data_frame()`