--- a/R/type-data-frame.R
+++ b/R/type-data-frame.R
@@ -16,9 +16,15 @@
 #' @examples
 #' new_data_frame(list(x = 1:10, y = 10:1))
 new_data_frame <- function(x = list(), n = NULL, ..., class = character()) {
-  stopifnot(is.list(x))
-  n <- n %||% df_size(x)
-  stopifnot(is.integer(n), length(n) == 1L)
+  if (!is.list(x)) {
+    abort("`x` must be a list.")
+  }
+
+  if (is.null(n)) {
+    n <- df_size(x)
+  } else if (!is.integer(n) || length(n) != 1L) {
+    abort("`n` must be an integer of size 1.")
+  }
 
   # names() should always be a character vector, but we can't enforce that
   # because as.data.frame() returns a data frame with NULL names to indicate
@@ -27,12 +33,15 @@ new_data_frame <- function(x = list(), n = NULL, ..., class = character()) {
     names(x) <- character()
   }
 
-  structure(
-    x,
+  new_attributes <- list(
     ...,
     class = c(class, "data.frame"),
     row.names = .set_row_names(n)
   )
+
+  attributes(x) <- c(attributes(x), new_attributes)
+
+  x
 }
 
 # Light weight constructor used for tests - avoids having to repeatedly do