Helps with #535 (maybe closes?)

`new_data_frame()` is _relatively_ slow, and since it can potentially be called many times through the use of `vec_proxy.POSIXlt()`, it can sometimes hurt performance. I've removed the usage of `stopifnot()` and `structure()` to get better performance.

``` r
library(vctrs)

ct <- new_datetime(0)
lt <- as.POSIXlt(ct)
lt_unclassed <- unclass(lt)
```

`new_data_frame()` on a list is now faster:

```r
# before
bench::mark(new_data_frame(lt_unclassed), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression                       min  median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                   <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl>
#> 1 new_data_frame(lt_unclassed)  11.6µs  16.9µs    58050.    24.8KB     63.9

# after
bench::mark(new_data_frame(lt_unclassed), iterations = 100000)
#> # A tibble: 1 x 6
#>   expression                       min  median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                   <bch:t> <bch:t>     <dbl> <bch:byt>    <dbl>
#> 1 new_data_frame(lt_unclassed)  4.36µs  6.05µs   162583.    30.2KB     79.7
```

Resulting in better performance when using `vec_size()` on a POSIXlt, which has to get the `vec_proxy.POSIXlt()`, which calls `new_data_frame()`.

```r
# before
bench::mark(vec_size(ct), vec_size(lt), iterations = 100000)
#> # A tibble: 2 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_size(ct)    522ns    645ns  1442711.     6.3KB     14.4
#> 2 vec_size(lt)     14µs   18.7µs    54743.    4.68KB     23.5

# after
bench::mark(vec_size(ct), vec_size(lt), iterations = 100000)
#> # A tibble: 2 x 6
#>   expression        min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>   <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_size(ct)    501ns    572ns  1370221.     6.3KB     13.7
#> 2 vec_size(lt)   5.79µs   7.08µs   134353.    4.68KB     18.8
```